split the code in cache exchange into separate files
- optimistic updates and manual updates are now in separate files
- generic cache helpers are in a separate file
