+++ project-base/storefront/urql/cache/cacheExchange.ts
@@ -1,0 +1,87 @@
+import { optimisticUpdates } from './optimistic';
+import { cacheUpdates } from './updates';
+import { cacheExchange, Data } from '@urql/exchange-graphcache';
+import { IntrospectionQuery } from 'graphql';
+import schema from 'schema.graphql.json';
+
+const keyNull = () => null;
+const keyUuid = (data: Data) => data.uuid as string | null;
+const keyName = (data: Data) => data.name as string | null;
+const keyCode = (data: Data) => data.code as string | null;
+const keyUrl = (data: Data) => data.url as string | null;
+
+export const cache = cacheExchange({
+    schema: schema as unknown as IntrospectionQuery,
+    keys: {
+        AdditionalSize: keyUrl,
+        Advert: keyUuid,
+        AdvertCode: keyUuid,
+        AdvertImage: keyUuid,
+        AdvertPosition: (data) => data.positionName as string | null,
+        ArticleSite: keyUuid,
+        Availability: keyName,
+        BlogArticle: keyUuid,
+        BlogCategory: keyUuid,
+        Brand: keyUuid,
+        BrandFilterOption: keyNull,
+        Cart: keyUuid,
+        CartItem: keyUuid,
+        CartItemModificationsResult: keyNull,
+        CartModificationsResult: keyNull,
+        CartMultipleAddedProductModificationsResult: keyNull,
+        CartPaymentModificationsResult: keyNull,
+        CartTransportModificationsResult: keyNull,
+        CartPromoCodeModificationsResult: keyNull,
+        Category: keyUuid,
+        CompanyCustomerUser: keyUuid,
+        Country: keyCode,
+        CustomerUser: keyUuid,
+        DeliveryAddress: keyUuid,
+        File: keyNull,
+        Flag: keyUuid,
+        FlagFilterOption: keyNull,
+        GoPayPaymentMethod: (data) => data.identifier as string | null,
+        Image: keyNull,
+        ImageSize: keyUrl,
+        Link: keyNull,
+        MainVariant: keyUuid,
+        NavigationItem: keyNull,
+        NavigationItemCategoriesByColumns: keyNull,
+        NewsletterSubscriber: keyNull,
+        NotificationBar: keyNull,
+        Order: keyUuid,
+        OrderItem: keyNull,
+        OpeningHours: keyUuid,
+        OpeningHoursOfDay: keyUuid,
+        Parameter: keyNull,
+        ParameterCheckboxFilterOption: keyNull,
+        ParameterSliderFilterOption: keyNull,
+        ParameterColorFilterOption: keyNull,
+        ParameterValueColorFilterOption: keyNull,
+        ParameterValue: keyUuid,
+        ParameterValueFilterOption: keyNull,
+        Payment: keyUuid,
+        PersonalData: keyNull,
+        PersonalDataPage: keyNull,
+        Price: keyNull,
+        PricingSetting: keyNull,
+        Product: keyUuid,
+        ProductFilterOptions: keyNull,
+        ProductPrice: keyNull,
+        RegularCustomerUser: keyUuid,
+        RegularProduct: keyUuid,
+        SeoSetting: keyNull,
+        SeoPage: keyNull,
+        Settings: keyNull,
+        SliderItem: keyUuid,
+        Store: keyUuid,
+        StoreAvailability: keyNull,
+        Transport: keyUuid,
+        TransportType: keyCode,
+        Unit: keyName,
+        Variant: keyUuid,
+        ProductList: keyUuid,
+    },
+    updates: cacheUpdates,
+    optimistic: optimisticUpdates,
+});
